requirements.txt
langchain==0.0.184
python-dotenv==1.0.0
youtube_transcript_api==0.6.1
